VotingApp/vote_app/requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
protobuf==3.13.0
PyMySQL==0.10.1
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
